pyproject.toml
# SPDX-FileCopyrightText: 2021-2024 Helmholtz-Zentrum hereon GmbH
#
# SPDX-License-Identifier: CC0-1.0

[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools >= 61.0', 'versioneer[toml]']

[project]
name = "psy-maps"
dynamic = ["version"]
description = "Psyplot plugin for visualization on a map"

readme = "README.rst"
keywords = [
    "visualization",
    "psyplot",
    "netcdf",
    "matplotlib",
    "cartopy",
    "maps",
    "georeferenced",
    ]

authors = [
    { name = 'Philipp S. Sommer', email = 'philipp.sommer@hereon.de' },
]
maintainers = [
    { name = 'Philipp S. Sommer', email = 'philipp.sommer@hereon.de' },
]
license = { text = 'LGPL-3.0-only' }

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]

requires-python = '>= 3.9'
dependencies = [
    "psyplot>=1.3.0",
    # add your dependencies here
    "psy-simple>=1.3.0",
    "cartopy",
    "scipy",
]

[project.urls]
Homepage = 'https://codebase.helmholtz.cloud/psyplot/psy-maps'
Documentation = "https://psyplot.github.io/psy-maps"
Source = "https://codebase.helmholtz.cloud/psyplot/psy-maps"
Tracker = "https://codebase.helmholtz.cloud/psyplot/psy-maps/issues/"


[project.entry-points."psyplot"]
plugin = "psy_maps.plugin"
patches = "psy_maps.plugin:patches"

[project.optional-dependencies]
testsite = [
    "tox",
    "isort==5.12.0",
    "black==23.1.0",
    "blackdoc==0.3.8",
    "flake8==6.0.0",
    "pre-commit",
    "mypy",
    "pytest-cov",
    "reuse",
    "cffconvert",
    "pytest-xdist",
    "dask",
    "netCDF4",
    "seaborn",
    "psyplot_gui",
    "pytest-qt",
]
docs = [
    "autodocsumm",
    "sphinx-rtd-theme",
    "hereon-netcdf-sphinxext",
    "sphinx-design",

]
dev = [
    "psy-maps[testsite]",
    "psy-maps[docs]",
    "PyYAML",
    "types-PyYAML",
]


[tool.mypy]
ignore_missing_imports = true

[tool.setuptools]
zip-safe = false
license-files = ["LICENSES/*"]

[tool.setuptools.package-data]
psy_maps = ["py.typed"]

[tool.setuptools.packages.find]
namespaces = false
exclude = [
    'docs',
    'tests*',
    'examples'
]

[tool.pytest.ini_options]
addopts = '-v'

[tool.versioneer]
VCS = 'git'
style = 'pep440'
versionfile_source = 'psy_maps/_version.py'
versionfile_build = 'psy_maps/_version.py'
tag_prefix = 'v'
parentdir_prefix = 'psy-maps-'

[tool.isort]
profile = "black"
line_length = 79
src_paths = ["psy_maps"]
float_to_top = true
known_first_party = "psy_maps"

[tool.black]
line-length = 79
target-version = ['py39']

[tool.coverage.run]
omit = ["psy_maps/_version.py"]
